requirements.txt
boto3>=1.26.0
botocore>=1.29.0
fastapi>=0.104.0
flask>=3.0.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
dataclasses-json>=0.5.14
PyJWT>=2.8.0
cryptography>=41.0.0
mangum>=0.17.0
